Faster computation of scalar products
Based on info about rational classes.
In internal characters of a UTable we store only the value on one
conjugacy class in any rational class. The other values of generalized
characters are computed by Galois conjugation.

Furthermore we determine the conductor c of the character field of a
class. Instead of storing values on this class as cyclotomics in CF(c)
we store them as Phi(c) integer coefficients when written as linear
combinations of 1, E(c), .., E(c)^(Phi(c)-1).

This way taking Z-linear combinations of generalized characters
(stored as integer vectors) becomes very fast.

Also there is a specialized function which computes scalar products of
characters stored this way. This uses a list of integers constructed
from RationalClassesInfo(G).